Added a Flag type to the CLI, plus coloured text. Two things have been added:
1. There is now a Flag type that resolves into a boolean of whether it was mentioned or not. This is useful for adding configuration options to any CLI commands.
2. Warnings and errors now have coloured text. I'll probably find a way to raise an error with coloured text, but that's for another time.

I'm liking how this is coming along. It's really cool to see all the things I can do with this.

Once I get overloads working correctly, I'll create a built-in help command that you can use for help on any command in the CLI, whether it's user-made or built-in. It'll most likely read from __doc__ and maybe include a summary in the decorator, but who knows?

We'll see what I come up with.

Toodeloo!